Getting back with FizzBuzz
Resuming Clojure studies with FizzBuzz. Getting to know how cond is
useful in this situation and remembering that every single operator is a
function and should, therefore, be used as such:

(mod n 3) == 0 ; Compiles but does not work as expected
(= 0 (mod n 3)) ; This is the correct form

Also learned that cond does not require :else to be described, nil will
be returned when no conditions are met.